Bump setuptools to 80.3.*
