+++ stubs/setuptools/METADATA.toml
@@ -1,4 +1,4 @@
-version = "80.1.*"
+version = "80.3.*"
 upstream_repository = "https://github.com/pypa/setuptools"
 extra_description = """\
 Given that `pkg_resources` is typed since `setuptools >= 71.1`, \
