+++ stubs/setuptools/setuptools/_scripts.pyi
@@ -1,0 +1,46 @@
+from _typeshed import Incomplete, Unused
+from collections.abc import Iterable, Iterator
+from typing import ClassVar, Literal, TypedDict
+from typing_extensions import Self
+
+class _SplitArgs(TypedDict, total=False):
+    comments: bool
+    posix: bool
+
+class CommandSpec(list[str]):
+    options: list[str]
+    split_args: ClassVar[_SplitArgs]
+    @classmethod
+    def best(cls) -> type[CommandSpec]: ...
+    @classmethod
+    def from_param(cls, param: Self | str | Iterable[str] | None) -> Self: ...
+    @classmethod
+    def from_environment(cls) -> CommandSpec: ...
+    @classmethod
+    def from_string(cls, string: str) -> CommandSpec: ...
+    def install_options(self, script_text: str) -> None: ...
+    def as_header(self) -> str: ...
+
+class WindowsCommandSpec(CommandSpec): ...
+
+class ScriptWriter:
+    template: ClassVar[str]
+    command_spec_class: ClassVar[type[CommandSpec]]
+    @classmethod
+    def get_args(cls, dist, header: Incomplete | None = None) -> Iterator[tuple[str, str]]: ...
+    @classmethod
+    def best(cls) -> type[ScriptWriter]: ...
+    @classmethod
+    def get_header(cls, script_text: str = "", executable: str | CommandSpec | Iterable[str] | None = None) -> str: ...
+
+class WindowsScriptWriter(ScriptWriter):
+    command_spec_class: ClassVar[type[WindowsCommandSpec]]
+    @classmethod
+    def best(cls) -> type[WindowsScriptWriter]: ...
+
+class WindowsExecutableLauncherWriter(WindowsScriptWriter): ...
+
+def get_win_launcher(type: Literal["cli", "gui"]) -> bytes: ...
+def load_launcher_manifest(name: Unused) -> str: ...
+def is_64bit() -> bool: ...
+def isascii(s: str) -> bool: ...
